# openapi/spec.yaml
openapi: 3.1.0
info:
  title: Energy Control API
  version: 0.1.0
  description: Intelligent grid setpoint controller for Victron solar systems
paths:
...
  /api/verschattung/config:
    get:
      operationId: getVerschattungConfig
      responses:
        '200':
          description: Config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerschattungConfig'
    put:
      operationId: putVerschattungConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerschattungConfig'
      responses:
        '200':
          description: Updated config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerschattungConfig'
  /api/verschattung/cover/{id}/position:
    put:
      operationId: putVerschattungCoverPosition
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                position:
                  type: number
              required: [position]
      responses:
        '204':
          description: Sent
  /api/verschattung/cover/{id}/auto:
    post:
      operationId: postVerschattungCoverAuto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Override released
components:
  schemas:
...
    VerschattungConfig:
      type: object
      additionalProperties: false
      required: [zones, pvThreshold, indoorTempThresholdC, hysteresisIndoorTempC, hysteresisPvFactor, hysteresisPvDurationMinutes, summerModeMonths]
      properties:
        zones:
          type: object
          required: [ost, sued, west]
          properties:
            ost:  { $ref: '#/components/schemas/VerschattungZoneTunables' }
            sued: { $ref: '#/components/schemas/VerschattungZoneTunables' }
            west: { $ref: '#/components/schemas/VerschattungZoneTunables' }
        pvThreshold:
          type: object
          required: [peakWp, factor, floorW]
          properties:
            peakWp: { type: number }
            factor: { type: number }
            floorW: { type: number }
        indoorTempThresholdC: { type: number }
        hysteresisIndoorTempC: { type: number }
        hysteresisPvFactor: { type: number }
        hysteresisPvDurationMinutes: { type: number }
        summerModeMonths:
          type: array
          items: { type: integer, minimum: 1, maximum: 12 }
    VerschattungZoneTunables:
      type: object
      required: [closePosition]
      properties:
        closePosition: { type: number, minimum: 0, maximum: 100 }